Updated WV name to WorkoutView
